requirements.txt
fastapi==0.115.12
jinja2==3.1.6
pip-chill==1.0.3
python-dotenv==1.0.1
uvicorn==0.34.0
pydantic-settings==2.8.1
langchain==0.3.22
pypdf==5.4.0
docx2txt==0.9
sentence-transformers==4.0.1
pysqlite3-binary==0.5.4
chromadb==0.6.3
python-multipart==0.0.20
celery==5.4.0
